% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_rain.R
\docType{data}
\name{ad_rain}
\alias{ad_rain}
\title{Data frame of daily rainfall data from NOAA NCDC from 2017 to 2023}
\format{
A \code{data.frame}
}
\usage{
ad_rain
}
\description{
Data frame of daily rainfall data from NOAA NCDC from 2017 to 2023
}
\details{
Used for estimating atmospheric deposition and created using the \code{\link{util_ad_getrain}} function. The data frame contains the following columns:

\itemize{
\item \code{stationid}: Character string for the station id
\item \code{date}: Date for the observation
\item \code{Year}: Numeric value for the year of the observation
\item \code{Month}: Numeric value for the month of the observation
\item \code{Day}: Numeric value for the day of the observation
\item \code{rainfall}: Numeric value for the amount of rainfall in inches
}
}
\examples{
ad_rain
}
\seealso{
\code{\link{util_ad_getrain}}
}
\keyword{datasets}
