% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allflo.R
\docType{data}
\name{allflo}
\alias{allflo}
\title{Data frame of all flow data used in \code{\link{anlz_nps_gaged}} and \code{\link{anlz_nps_ungaged}}}
\format{
A \code{data.frame} of monthly mean daily flow data for select basins
}
\usage{
allflo
}
\description{
Data frame of all flow data used in \code{\link{anlz_nps_gaged}} and \code{\link{anlz_nps_ungaged}}
}
\details{
Monthly flow data at select stations used for estimating non-point source gaged and ungaged loads.  Created using the \code{\link{util_nps_getflow}} function. Includes data from the USGS API using \code{\link{util_nps_getusgsflow}} and from external sources using \code{\link{util_nps_getextflow}}. The data frame contains the following columns:

\itemize{
\item \code{basin}: Character string for the basin or gauge location
\item \code{yr}: Year of the observation
\item \code{mo}: Month of the observation
\item \code{flow_cfs}: Numeric value for the average daily flow in cubic feet per second (cfs)
}
}
\examples{
allflo
}
\seealso{
\code{\link{util_nps_getusgsflow}}, \code{\link{util_nps_getextflow}}, \code{\link{util_nps_getflow}}
}
\keyword{datasets}
