% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allwq.R
\docType{data}
\name{allwq}
\alias{allwq}
\title{Data frame of all water quality data used in \code{\link{anlz_nps_gaged}}}
\format{
A \code{data.frame} of monthly mean water quality data for select stations
}
\usage{
allwq
}
\description{
Data frame of all water quality data used in \code{\link{anlz_nps_gaged}}
}
\details{
Monthly water quality data for select stations used for estimating non-point source gaged loads.  Created using the \code{\link{util_nps_getwq}} function. Includes data from Manatee, Pinellas, and Hillsborough (EPCHC) counties. The data frame contains the following columns:

\itemize{
\item \code{basin}: Character string for the basin or station location
\item \code{yr}: Year of the observation
\item \code{mo}: Month of the observation
\item \code{tn_mgl}: Numeric value for Total Nitrogen in mg/L
\item \code{tp_mgl}: Numeric value for Total Phosphorus in mg/L
\item \code{tss_mgl}: Numeric value for Total Suspended Solids in mg/L
\item \code{bod_mgl}: Numeric value for Biochemical Oxygen Demand in mg/L
}
}
\examples{
allwq
}
\seealso{
\code{\link{util_nps_getwq}}
}
\keyword{datasets}
