% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_ad.R
\name{anlz_ad}
\alias{anlz_ad}
\title{Calculate AD loads and summarize}
\usage{
anlz_ad(
  ad_rain,
  vernafl,
  summ = c("segment", "all"),
  summtime = c("month", "year")
)
}
\arguments{
\item{vernafl}{character vector of file path to Verna Wellfield atmospheric concentration data}

\item{summ}{chr string indicating how the returned data are summarized, see details}

\item{summtime}{chr string indicating how the returned data are summarized temporally (month or year), see details}

\item{rain}{data frame of daily rainfall data from NOAA NCDC, obtained using \code{\link{util_ad_getrain}}}
}
\value{
A data frame with nitrogen and phosphorus loads in tons/month, hydrologic load in million m3/month, and segment, year, and month as columns if \code{summ = 'segment'} and \code{summtime = 'month'}. Total load to all segments can be returned if \code{summ = 'all'} and annual summaries can be returned if \code{summtime = 'year'}.  In the latter case, loads are the sum of monthly estimates such that output is tons/yr for TN and TP and as million m3/yr for hydrologic load.
}
\description{
Calculate AD loads and summarize
}
\details{
Loading from atmospheric deposition (AD) for bay segments in the Tampa Bay watershed are calculated using rainfall data and atmospheric concentration data from the Verna Wellfield site.  Rainfall data must be obtained using the \code{\link{util_ad_getrain}} function before calculating loads.  For convenience, daily rainfall data from 2017 to 2023 at sites in the watershed are included with the package in the \code{\link{ad_rain}} object.  The Verna Wellfield data must also be obtained from \url{https://nadp.slh.wisc.edu/sites/ntn-FL41/} as monthly observations. This file is also included with the package and can be found using \code{\link{system.file}} as in the examples below. Internally, the Verna data are converted to total nitrogen and total phosphorus from ammonium and nitrate concentration data (see \code{\link{util_ad_prepverna}} for additional information).

The function first estimates the total hydrologic load for each bay segment using daily estimates of rainfall at NWIS NCDC sites in the watershed.  This is done as a weighted mean of rainfall at the measured sites relative to grid locations in each sub-watershed for the bay segments.  The weights are based on distance of the grid cells from the closest site as inverse distance squared.  Total hydrologic load for a sub-watershed is then estimated by converting inches/month  to m3/month using the area of each watershed.  The distance data and sub-watershed areas are contained in the \code{\link{ad_distance}} file included with the package.

The total nitrogen and phosphorus loads are then estimated for each bay segment by multiplying the total hydrologic load by the total nitrogen and phosphorus concentrations in the Verna data.  The loading calculations also include a wet/dry deposition conversion factor to account for differences in loading during the rainy and dry seasons.
}
\examples{
vernafl <- system.file('extdata/verna-raw.csv', package = 'tbeploads')
data(ad_rain)
anlz_ad(ad_rain, vernafl)
}
\seealso{
\code{\link{util_ad_getrain}}, \code{\link{util_ad_prepverna}}
}
