% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_ips.R
\name{anlz_ips}
\alias{anlz_ips}
\title{Calculate IPS loads and summarize}
\usage{
anlz_ips(
  fls,
  summ = c("entity", "facility", "segment", "all"),
  summtime = c("month", "year")
)
}
\arguments{
\item{fls}{vector of file paths to raw entity data, one to many}

\item{summ}{chr string indicating how the returned data are summarized, see details}

\item{summtime}{chr string indicating how the returned data are summarized temporally (month or year), see details}
}
\value{
data frame with loading data for TP, TN, TSS, and BOD as tons per month/year and hydro load as million cubic meters per month/year
}
\description{
Calculate IPS loads and summarize
}
\details{
Input data files in \code{fls} are first processed by \code{\link{anlz_ips_facility}} to calculate IPS loads for each facility and outfall.  The data are summarized differently based on the \code{summ} and \code{summtime} arguments.  All loading data are summed based on these arguments, e.g., by bay segment (\code{summ = 'segment'}) and year (\code{summtime = 'year'}).
}
\examples{
fls <- list.files(system.file('extdata/', package = 'tbeploads'),
  pattern = 'ps_ind', full.names = TRUE)
anlz_ips(fls)
}
\seealso{
\code{\link{anlz_ips_facility}}
}
