% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_nps.R
\name{anlz_nps}
\alias{anlz_nps}
\title{Calculate non-point source (NPS) loads for Tampa Bay}
\usage{
anlz_nps(
  yrrng = c("2021-01-01", "2023-12-31"),
  tbbase,
  rain,
  mancopth,
  pincopth,
  lakemanpth,
  tampabypth,
  bellshlpth,
  vernafl,
  summ = c("segment", "all", "lu"),
  summtime = c("month", "year"),
  verbose = T
)
}
\arguments{
\item{yrrng}{A vector of two dates in 'YYYY-MM-DD' format, specifying the date range to retrieve flow data. Default is from '2021-01-01' to '2023-12-31'.}

\item{tbbase}{data frame containing polygon areas for the combined data layer of bay segment, basin, jurisdiction, land use data, and soils, see details}

\item{rain}{data frame of rainfall data, see details}

\item{mancopth}{character, path to the Manatee County water quality data file, see details}

\item{pincopth}{character, path to the Pinellas County water quality data file, see details}

\item{lakemanpth}{character, path to the file containing the Lake Manatee flow data, see details}

\item{tampabypth}{character, path to the file containing the Tampa Bypass flow data, see details}

\item{bellshlpth}{character, path to the file containing the Bell shoals data, see details}

\item{vernafl}{character vector of file path to Verna Wellfield atmospheric concentration data}

\item{summ}{character indicating how to summarrize results, by land use (\code{'lu'}, ungaged basins only), by bay segment (\code{'segment'}, default), or all segments combined (\code{'all'})}

\item{summtime}{chr string indicating how the returned data are summarized temporally (month or year), see details}

\item{verbose}{logical indicating whether to print verbose output}
}
\value{
A data frame of non-point source loads for Tampa Bay, including columns for year, month, basin, bay segment, basin area (hectares), and loads for water, total nitrogen (TN), total phosphorus (TP), total suspended solids (TSS), and biochemical oxygen demand (BOD).
}
\description{
Calculate non-point source (NPS) loads for Tampa Bay
}
\details{
The function estimates non-point source (NPS) loads for Tampa Bay by combining ungaged and gaged NPS loads. Ungaged loads are estimated using rainfall, flow, event mean concentration, land use, and soils data, while gaged loads are estimated using water quality data and flow data. The function also incorporates atmospheric concentration data from the Verna Wellfield site.

The data are summarized differently based on the \code{summ} and \code{summtime} arguments.  All loading data are summed based on these arguments, e.g., by bay segment (\code{summ = 'segment'}) and year (\code{summtime = 'year'}). Using \code{summ = 'lu'} will return only ungaged loads summarized by land use.

The following functions are used internally and are provided here for reference on the components used in the calculations:

\itemize{
\item \code{\link{anlz_nps_ungaged}}: Estimates ungaged NPS loads.
\item \code{\link{anlz_nps_gaged}}: Estimates gaged NPS loads.
\item \code{\link{util_prepverna}}: Prepares and fills missing data with five-year means for the Verna Wellfield site data.
\item \code{\link{util_nps_getflow}}: Gets flow estimates for NPS gaged and ungaged calculations.
\item \code{\link{util_nps_getusgsflow}}: Gets USGS flow data for NPS calculations, used in \code{\link{util_nps_getflow}}.
\item \code{\link{util_nps_getextflow}}: Gets external flow data and contacts (Lake Manatee, Tampa Bypass, and Bell Shoals), used in \code{\link{util_nps_getflow}}.
\item \code{\link{util_nps_getwq}}: Gets water quality data for NPS gaged calculations (for \code{mancopth}, \code{pincopth}, and Hillsborough County data).
\item \code{\link{util_getrain}}: Gets rainfall data for NPS ungaged calculations, as provided in \code{\link{rain}}.
\item \code{\link{util_nps_tbbase}}: Gets the combined spatial data layer for NPS ungaged calculations, as provided in \code{\link{tbbase}}.
}
}
\examples{
\dontrun{
data(tbbase)
data(rain)
mancopth <- system.file('extdata/nps_wq_manco.txt', package = 'tbeploads')
pincopth <- system.file('extdata/nps_wq_pinco.txt', package = 'tbeploads')
lakemanpth <- system.file('extdata/nps_extflow_lakemanatee.xlsx', package = 'tbeploads')
tampabypth <- system.file('extdata/nps_extflow_tampabypass.xlsx', package = 'tbeploads')
bellshlpth <- system.file('extdata/nps_extflow_bellshoals.xls', package = 'tbeploads')
vernafl <- system.file('extdata/verna-raw.csv', package = 'tbeploads')

anlz_nps(yrrng = c('2021-01-01', '2023-12-31'), tbbase, rain, mancopth, pincopth,
         lakemanpth, tampabypth, bellshlpth, vernafl, verbose = TRUE)
}
}
