% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_nps_ungaged.R
\name{anlz_nps_ungaged}
\alias{anlz_nps_ungaged}
\title{Estimated non-point source (NPS) ungaged loads}
\usage{
anlz_nps_ungaged(tbbase, rain, lakemanpth, tampabypth, bellshlpth)
}
\arguments{
\item{tbbase}{data frame containing polygon areas for the combined data layer of bay segment, basin, jurisdiction, land use data, and soils, see details}

\item{rain}{data frame of rainfall data, see details}

\item{lakemanpth}{character, path to the file containing the Lake Manatee flow data, see details}

\item{tampabypth}{character, path to the file containing the Tampa Bypass flow data, see details}

\item{bellshlpth}{character, path to the file containing the Bell shoals data, see details}
}
\value{
A data frame
}
\description{
Estimated non-point source (NPS) ungaged loads
}
\details{
Requires the following inputs:

\itemize{
\item \code{tbbase}: A data frame containing polygon areas for the combined data layer of bay segment, basin, jurisdiction, land use data, and soils, Stored as \code{\link{tbbase}} or created (takes an hour or so) with \code{\link{util_nps_gettbbase}}.
\item \code{rain}: A data frame of rainfall data see \code{\link{rain}}.
\item \code{lakemanpth}: character, path to the file containing the Lake Manatee flow data. See \code{\link{util_nps_getextflow}}.
\item \code{tampabypth}: character, path to the file containing the Tampa Bypass flow data. See \code{\link{util_nps_getextflow}}.
\item \code{bellshlpth}: character, path to the file containing the Bell shoals data. See \code{\link{util_nps_getextflow}}.
}

USGS gaged flows are also used as returned by \code{\link{util_nps_getusgsflow}}.
}
\examples{
# external flow sources
lakemanpth <- system.file('extdata/nps_extflow_lakemanatee.xlsx', package = 'tbeploads')
tampabypth <- system.file('extdata/nps_extflow_tampabypass.xlsx', package = 'tbeploads')
bellshlpth <- system.file('extdata/nps_extflow_bellshoals.xls', package = 'tbeploads')

anlz_nps_ungaged(tbbase, rain, lakemanpth, tampabypth, bellshlpth)
}
