% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contwet.R
\docType{data}
\name{contwet}
\alias{contwet}
\title{Upper Floridan Aquifer potentiometric surface raster, wet season 2022}
\format{
A \code{PackedSpatRaster} (see \code{\link[terra]{wrap}})
representing a 1-mile resolution grid of potentiometric head (ft above
MSL) for the wet season. Unwrap with \code{terra::unwrap(contwet)} to
obtain a \code{\link[terra]{SpatRaster}}.
}
\usage{
contwet
}
\description{
Upper Floridan Aquifer potentiometric surface raster, wet season 2022
}
\details{
Interpolated from Upper Floridan Aquifer potentiometric surface contour
lines (September 2022) downloaded from the FDEP / Florida Geological Survey
ArcGIS REST service using \code{\link{util_gw_getcontour}}. The spatial
extent covers the Tampa Bay watershed (\code{\link{tbfullshed}}) buffered
outward by 40 miles. Interpolation used inverse distance weighting (IDW,
5-mile radius, power = 2) followed by five passes of a 3x3 focal mean gap
fill. Projection is NAD83(2011) / Florida GDL Albers (ftUS), CRS 6443.

Dry season equivalent is \code{\link{contdry}}.
}
\examples{
\dontrun{
pot_wet <- util_gw_getcontour("wet", 2022)
contwet <- terra::wrap(pot_wet)
save(contwet, file = "data/contwet.RData", compress = "xz")
}
terra::unwrap(contwet)
}
\keyword{datasets}
