% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbasing.R
\docType{data}
\name{dbasing}
\alias{dbasing}
\title{Basin information for coastal subbasin codes}
\format{
A \code{data.frame}
}
\usage{
dbasing
}
\description{
Basin information for coastal subbasin codes
}
\details{
Used for domestic point source summaries, bay segments are as follows:

\itemize{
\item 1: Old Tampa Bay
\item 2: Hillsborough Bay
\item 3: Middle Tampa Bay
\item 4: Lower Tampa Bay
\item 5: Boca Ciega Bay
\item 6: Terra Ceia Bay
\item 7: Manatee River
\item 55: Boca Ciega Bay South
}

See "data-raw/dbasing.R" for creation.
}
\examples{
dbasing
}
\keyword{datasets}
