% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpsfac.R
\docType{data}
\name{dpsfac}
\alias{dpsfac}
\title{Domestic point source facilities}
\format{
A \code{data.frame}

#examples
dpsfac
}
\usage{
dpsfac
}
\description{
Domestic point source facilities
}
\keyword{datasets}
