% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emc.R
\docType{data}
\name{emc}
\alias{emc}
\title{Event Mean Concentration (EMC) data for CLUCSID in Tampa Bay}
\format{
A \code{data.frame}
}
\usage{
emc
}
\description{
Event Mean Concentration (EMC) data for CLUCSID in Tampa Bay
}
\details{
Used for non-point source (NPS) ungaged estimates summaries, bay segments are as follows:

Values are grouped by CLUCSID and include mean TN, TP, TSS, and BOD.

See "data-raw/emc.R" for creation.
}
\examples{
emc
}
\keyword{datasets}
