% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facilities.R
\docType{data}
\name{facilities}
\alias{facilities}
\title{Domestic and industrial point source facilities}
\format{
A \code{data.frame}

#examples
facilities
}
\usage{
facilities
}
\description{
Domestic and industrial point source facilities
}
\keyword{datasets}
