% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsfac.R
\docType{data}
\name{ipsfac}
\alias{ipsfac}
\title{Industrial point source facilities}
\format{
A \code{data.frame}

#examples
ipsfac
}
\usage{
ipsfac
}
\description{
Industrial point source facilities
}
\keyword{datasets}
