% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbbase.R
\docType{data}
\name{tbbase}
\alias{tbbase}
\title{Combined spatial data required for non-point source (NPS) ungaged estimate}
\format{
A summarized data frame containing the union of all inputs showing major bay segment, sub-basin (basin), drainage feature (drnfeat), jurisdiction (entity), land use/land cover (FLUCCSCODE), CLUCSID, IMPROVED, hydrologic group (hydgrp), and area in hectures.  These represent all relevant spatial combinations in the Tampa Bay watershed.

See "data-raw/tbbase.R" for creation.
}
\usage{
tbbase
}
\description{
Combined spatial data required for non-point source (NPS) ungaged estimate
}
\examples{
tbbase
}
\keyword{datasets}
