% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbdbasin.R
\docType{data}
\name{tbdbasin}
\alias{tbdbasin}
\title{Simple feature polygons major drainage basins in the Tampa Bay Estuary Program boundary}
\format{
A \code{\link[sf]{sf}} object
}
\usage{
tbdbasin
}
\description{
Simple feature polygons major drainage basins in the Tampa Bay Estuary Program boundary
}
\details{
Used for estimating ungaged non-point source (NPS) loads. The data includes the following columns.

\itemize{
\item \code{basin}: Numeric value for the basin
\item \code{drnfeat}: Numeric for the drainage feature
\item \code{geometry}: The geometry column
}

Projection is NAD83(2011) / Florida West (ftUS), CRS 6443.
}
\examples{
\dontrun{
prj <- 6443

tbdbasin <- sf::st_read("./data-raw/TBEP/gis/TBEP_dBasins_FIPS0902_Projection.shp") |>
  sf::st_transform(prj) |>
  sf::st_buffer(dist = 0) |>
  dplyr::group_by(NEWGAGE, DRNFEATURE) |>
  dplyr::summarise(geometry = sf::st_union(geometry), .groups = "drop") |>
  dplyr::select(
    basin = NEWGAGE,
    drnfeat = DRNFEATURE
  ) |>
  dplyr::arrange(basin, drnfeat)

save(tbdbasin, file = 'data/tbdbasin.RData', compress = 'xz')
}
tbdbasin
}
\keyword{datasets}
