% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbjuris.R
\docType{data}
\name{tbjuris}
\alias{tbjuris}
\title{Simple feature polygons of jurisdictional boundaries in the Tampa Bay Estuary Program boundary}
\format{
A \code{\link[sf]{sf}} object
}
\usage{
tbjuris
}
\description{
Simple feature polygons of jurisdictional boundaries in the Tampa Bay Estuary Program boundary
}
\details{
Used for estimating ungaged non-point source (NPS) loads. The data includes the following columns.

\itemize{
\item \code{entity}: Character for the entity name
\item \code{geometry}: The geometry column
}

Projection is NAD83(2011) / Florida West (ftUS), CRS 6443.
}
\examples{
\dontrun{
prj <- 6443

tbjuris <- sf::st_read("./data-raw/TBEP/gis/TB_Juris.shp") |>
  sf::st_transform(prj) |>
  sf::st_buffer(dist = 0) |>
  dplyr::rename(entity = NAME_FINAL) |>
  dplyr::select(entity)

save(tbjuris, file = 'data/tbjuris.RData', compress = 'xz')
}
tbjuris
}
\keyword{datasets}
