% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblu2020.R
\docType{data}
\name{tblu2020}
\alias{tblu2020}
\title{Simple feature polygons of 2020 land use in the Tampa Bay Estuary Program boundary}
\format{
A \code{\link[sf]{sf}} object
}
\usage{
tblu2020
}
\description{
Simple feature polygons of 2020 land use in the Tampa Bay Estuary Program boundary
}
\details{
Used for estimating ungaged non-point source (NPS) loads. The data includes the following columns.

\itemize{
\item \code{FLUCCSCODE}: Numeric value for the Florida Land Use, Cover and Forms Classification System (FLUCCS) code
\item \code{FLUCCSDESC}: Character describing the FLUCCS description
\item \code{geometry}: The geometry column
}

Projection is NAD83(2011) / Florida West (ftUS), CRS 6443.
}
\examples{
\dontrun{
# import from local source
prj <- 6443

tblu2020 <- sf::st_read("T:/05_GIS/SWFWMD/LULC_2020/LANDUSELANDCOVER2020.shp") |>
  sf::st_transform(prj) |>
  sf::st_intersection(tbfullshed) |>
  sf::st_buffer(dist = 0) |>
  dplyr::group_by(FLUCCSCODE, FLUCSDESC) |>
  dplyr::summarise() |>
  dplyr::ungroup()

# or use SWFWMD API
tblu2020 <- util_nps_getswfwmd('lulc2020')

save(tblu2020, file = 'data/tblu2020.RData', compress = 'xz')
}
}
\keyword{datasets}
