% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbsoil.R
\docType{data}
\name{tbsoil}
\alias{tbsoil}
\title{Simple feature polygons of soil data in the Tampa Bay Estuary Program boundary}
\format{
A \code{\link[sf]{sf}} object
}
\usage{
tbsoil
}
\description{
Simple feature polygons of soil data in the Tampa Bay Estuary Program boundary
}
\details{
Used for estimating ungaged non-point source (NPS) loads. The data includes the following columns.

\itemize{
\item \code{hydgrp}: Character for the hydrologic group (A, B, C, D, etc.) of the soil
\item \code{geometry}: The geometry column
}

Projection is NAD83(2011) / Florida West (ftUS), CRS 6443.
}
\examples{
\dontrun{
usdasoil <- st_read('T:/05_GIS/TBEP/TBLOADS/USDA_SSURGO_CLIP_FIPS0902.shp')
tbsoil <- usdasoil |>
  select(hydgrp = SDV_Hydr_1) |>
  group_by(hydgrp) |>
  summarise() |>
  st_transform(crs = 6443)
save(tbsoil, file = 'data/tbsoil.RData', compress = 'xz')
}
}
\keyword{datasets}
