% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgsflow.R
\docType{data}
\name{usgsflow}
\alias{usgsflow}
\title{Data frame of USGS stream flow data from the USGS NWIS database for 2021 to 2023}
\format{
A \code{data.frame}
}
\usage{
usgsflow
}
\description{
Data frame of USGS stream flow data from the USGS NWIS database for 2021 to 2023
}
\details{
Dadily flow data at select stations used for estimating non-point source gaged and ungaged loads.  Created using the \code{\link{util_getusgsflow}} function. The file is provided to reduce calls to the USGS API. The data frame contains the following columns:

\itemize{
\item \code{site_no}: Character string for the site number
\item \code{date}: Date for the observation
\item \code{flow_cfs}: Numeric value for the daily flow in cubic feet per second (cfs)
}
}
\examples{
usgsflow
}
\seealso{
\code{\link{util_getusgsflow}}
}
\keyword{datasets}
