% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ad_getrain.R
\name{util_ad_getrain}
\alias{util_ad_getrain}
\title{Get rainfall data for a given year at NOAA NCDC sites}
\usage{
util_ad_getrain(yrs, station = NULL, noaa_key, ntry = 5, quiet = FALSE)
}
\arguments{
\item{station}{numeric vector of station numbers to retrieve, see details}

\item{noaa_key}{character for the NOAA API key}

\item{ntry}{numeric for the number of times to try to download the data}

\item{quiet}{logical to print progress in the console}

\item{yr}{numeric vector for the years of data to retrieve}
}
\value{
a data frame with the following columns:
\itemize{

\item{stationid}{character, the station id}
\item{date}{Date, the date of the observation}
\item{Year}{numeric, the year of the observation}
\item{Month}{numeric, the month of the observation}
\item{Day}{numeric, the day of the observation}
\item{rainfall}{numeric, the amount of rainfall in inches}

}
}
\description{
Get rainfall data for a given year at NOAA NCDC sites
}
\details{
This function is used to retrieve a long-term record of rainfall for estimating AD loads.  It is used to create an input data file for load calculations and it is not used directly by any other functions due to download time.  A NOAA API key is required to use the function.

By default, rainfall data is retrieved for the stations 228, 478, 520, 940, 945, 1046, 1163, 1632, 1641, 2806, 3153, 3986, 4707, 5973, 6065, 6880, 7205, 7851, 7886, 8788, 8824, 9176, and 9401.
}
\examples{
noaa_key <- Sys.getenv('NOAA_KEY')
util_ad_getrain(2021, 228, noaa_key)
}
