% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ad_prepverna.R
\name{util_ad_prepverna}
\alias{util_ad_prepverna}
\title{Prep Verna Wellfield data for use in AD calculations}
\usage{
util_ad_prepverna(fl, fillmis = T)
}
\arguments{
\item{fl}{text string for the file path to the Verna Wellfield data}

\item{fillmis}{logical indicating whether to fill missing data with monthly means}
}
\value{
A data frame with total nitrogen and phosphorus estimates as mg/l for each year and month of the input data
}
\description{
Prep Verna Wellfield data for use in AD calculations
}
\details{
Raw data can be obtained from \url{https://nadp.slh.wisc.edu/sites/ntn-FL41/}.  Total nitrogen and phosphorus concentrations are estimated from ammonium and nitrate concentrations using the following relationships:

\deqn{TNConc = NH4 * 0.78 + NO3 * 0.23}
\deqn{TPConc = 0.01262 * TNConc + 0.00110}

The first equation corrects for the \% of ions in either that is N, and the second is a regression relationship between TBADS TN and TP, applied to Verna.
}
\examples{
fl <- list.files(system.file('extdata/', package = 'tbeploads'),
  pattern = 'verna-raw', full.names = TRUE)
util_ad_prepverna(fl)
}
