% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dps_addcol.R
\name{util_dps_addcol}
\alias{util_dps_addcol}
\title{Add column names for DPS reuse and end of pipe from raw entity data}
\usage{
util_dps_addcol(dat)
}
\arguments{
\item{dat}{data frame from raw entity data as \code{data.frame}}
}
\value{
Input data frame from \code{pth} as is if column names are correct, otherwise additional columns are added as needed.
}
\description{
Add column names for DPS reuse and end of pipe from raw entity data
}
\details{
The function checks for TN, TP, TSS, and BOD.  If any of these are missing, the columns are added with empty values including a column for units.  If BOD is missing but CBOD is present, the CBOD column is renamed to BOD.
}
\examples{
pth <- system.file('extdata/ps_dom_hillsco_falkenburg_2019.txt', package = 'tbeploads')
dat <- read.table(pth, skip = 0, sep = '\t', header = TRUE)
util_dps_addcol(dat)
}
