% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dps_checkfls.R
\name{util_dps_checkfls}
\alias{util_dps_checkfls}
\title{Create a data frame of formatting issues with DPS input files}
\usage{
util_dps_checkfls(pth)
}
\arguments{
\item{pth}{path to parent directory of raw entity data}
}
\value{
A \code{data.frame} with three columns indicating \code{name} for the file name, \code{chk} for the file issue, and \code{nms} for a concatenated string of column names for the file
}
\description{
Create a data frame of formatting issues with DPS input files
}
\details{
The \code{chk} column indicates the issue with the file and will indicate \code{"ok"} if no issues are found, \code{"read error"} if the file cannot be read, and \code{"check columns"} if the column names are not as expected.  Any file not showing \code{"ok"} should be checked for issues.

All files are checked with \code{\link{util_dps_checkuni}} if a file does not have a read error.
}
\examples{
pth <- system.file('extdata/ps_dom_hillsco_falkenburg_2019.txt', package = 'tbeploads')
pth <- dirname(pth)
util_dps_checkfls(pth)
}
