% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dps_checkuni.R
\name{util_dps_checkuni}
\alias{util_dps_checkuni}
\title{Check units for DPS reuse and end of pipe from raw entity data}
\usage{
util_dps_checkuni(dat)
}
\arguments{
\item{dat}{data frame from raw entity data as \code{data.frame}}
}
\value{
Input data frame from \code{pth} with relevant data, otherwise an error is returned if units are not correct.
}
\description{
Check units DPS reuse and end of pipe from raw entity data
}
\details{
Input data should include flow as million gallons per day, and concentration as mg/L.
}
\examples{
pth <- system.file('extdata/ps_dom_hillsco_falkenburg_2019.txt', package = 'tbeploads')
dat <- read.table(pth, skip = 0, sep = '\t', header = TRUE)
util_dps_checkuni(dat)
}
