% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dps_entfacseg.R
\name{util_dps_entfacseg}
\alias{util_dps_entfacseg}
\title{Get DPS entity, facility, and bay segment from file name}
\usage{
util_dps_entfacseg(pth, asdf = FALSE)
}
\arguments{
\item{pth}{path to raw entity data}

\item{asdf}{logical, if \code{TRUE} return as \code{data.frame}}
}
\value{
A list with entity, facility, and bay segment
}
\description{
Get DPS entity, facility, and bay segment from file name
}
\details{
Bay segment is an integer with values of 1, 2, 3, 4, 5, 6, 7, and 55 for Old Tampa Bay, Hillsborough Bay, Middle Tampa Bay, Lower Tampa Bay, Boca Ciega Bay, Terra Ceia Bay, Manatee River, and Boca Ciega Bay South, respectively.
}
\examples{
pth <- system.file('extdata/ps_dom_hillsco_falkenburg_2019.txt', package = 'tbeploads')
util_dps_entfacseg(pth)
}
