% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nps_fillmiswq.R
\name{util_nps_fillmiswq}
\alias{util_nps_fillmiswq}
\title{Fill in missing water quality values for non-point source (NPS) data}
\usage{
util_nps_fillmiswq(wq, yrrng = c("2021-01-01", "2023-12-31"))
}
\arguments{
\item{wq}{A data frame of water quality data returned by \code{util_nps_getwq}}

\item{yrrng}{A vector of two dates in 'YYYY-MM-DD' format, specifying the date range to retrieve flow data. Default is from '2021-01-01' to '2023-12-31'.}
}
\value{
Input data frame with missing data filled as described above.
}
\description{
Fill in missing water quality values for non-point source (NPS) data
}
\details{
Missing end date monthly values are filled with prior 5 year averages. Then, missing monthly values are linearly interpolated using \code{\link[zoo]{na.approx}}.
}
\examples{
\dontrun{
data(allwq)
wq <- util_nps_fillmiswq(allwq)
}
}
