% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nps_getextflow.R
\name{util_nps_getextflow}
\alias{util_nps_getextflow}
\title{Get external flow data not from USGS for NPS calculations}
\usage{
util_nps_getextflow(pth, loc, yrrng = c(2021, 2023))
}
\arguments{
\item{pth}{Path to the external Excel file}

\item{loc}{Location of the external flow data. Options are 'LMANATEE', 'TBYPASS', or '02301500'.}
}
\value{
A data frame of flow data for the location in \code{loc}
}
\description{
Get external flow data not from USGS for NPS calculations
}
\details{
This function retrieves and formats external flow data that cannot be obtained from the USGS API. The three required locations are Lake Manatee, Tampa Bypass Canal (s160), and the Alafia River Bell Shoals Bell Shoals.

External data can be obtained as follows:

\itemize{
\item LMANATEE: Lake Manatee flow for the Manatee River dam, from Manatee County Utilities, input flow is cfs.
\item TBYPASS: Tampa Bypass Canal flow from Tampa Bay Water.  Input flow is MGD and is converted to cfs (TBW contact is Cathleen Jonas, \href{mailto:cjonas@tampabaywater.org}{cjonas@tampabaywater.org}).
\item 02301500: Alafia River Bell Shoals flow data from SWFWMD WMIS Pumpage Reports for Permit 11794 or optionally from Tampa Bay Water reported withdrawals for Site 4626 (Cathleen Jonas, \href{mailto:cjonas@tampabaywater.org}{cjonas@tampabaywater.org}).  Input flow from latter is daily average converted to cfs.
}

System files are included in the package which can be updated annually.
}
\examples{
# lake manatee
pth <- system.file('extdata/nps_extflow_lakemanatee.xlsx', package = 'tbeploads')
extflo <- util_nps_getextflow(pth, loc = "LMANATEE")

# tampa bypass
pth <- system.file('extdata/nps_extflow_tampabypass.xlsx', package = 'tbeploads')
extflo <- util_nps_getextflow(pth, loc = "TBYPASS")

# bell shoals
pth <- system.file('extdata/nps_extflow_bellshoals.xls', package = 'tbeploads')
extflo <- util_nps_getextflow(pth, loc = "02301500")
}
