% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nps_landsoil.R
\name{util_nps_landsoil}
\alias{util_nps_landsoil}
\title{Utility function for non-point source (NPS) workflow to create land use and soil data}
\usage{
util_nps_landsoil(tbbase)
}
\arguments{
\item{tbbase}{Input data frame returned from \code{\link{util_nps_tbbase}}}
}
\value{
A data frame summarizing land use and soil by bay segment, sub-basin, drainage feature, CLUCSID, hydrologic group, and improved status.
}
\description{
Utility function for non-point source (NPS) workflow to create land use and soil data
}
\examples{
data(tbbase)
util_nps_landsoil(tbbase)
}
