% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nps_landsoilrc.R
\name{util_nps_landsoilrc}
\alias{util_nps_landsoilrc}
\title{Utility function to create non-point source (NPS) ungaged land use and soil runoff coefficients}
\usage{
util_nps_landsoilrc(tbbase, yrexp = c(2021:2023))
}
\arguments{
\item{tbbase}{Data frame returned from \code{\link{util_nps_tbbase}} containing land use and soil data.}

\item{yrexp}{Years to expand the data frame to include all months for each year.}
}
\value{
A data frame with land use (CLUCSID) and soil runoff coefficients by year and month.
}
\description{
Utility function to create non-point source (NPS) ungaged land use and soil runoff coefficients
}
\examples{
data(tbbase)

util_nps_landsoilrc(tbbase, yrexp = c(2021:2023))
}
