% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nps_lusumm.R
\name{util_nps_lusumm}
\alias{util_nps_lusumm}
\title{Summarize non-point source (NPS) ungaged loads by land use}
\usage{
util_nps_lusumm(
  dat,
  summ = c("basin", "segment", "all"),
  summtime = c("month", "year")
)
}
\arguments{
\item{dat}{Input data frame as an intermediate result from \code{\link{anlz_nps}}}

\item{summ}{chr string indicating how the returned data are summarized, see details}

\item{summtime}{chr string indicating how the returned data are summarized temporally (month or year), see details}
}
\value{
Data frame with summarized loading data based on user-supplied arguments
}
\description{
Summarize non-point source (NPS) ungaged loads by land use
}
\details{
The data are summarized differently based on the \code{summ} and \code{summtime} arguments.  All loading data are summed based on these arguments, e.g., by bay segment (\code{summ = 'segment'}) and year (\code{summtime = 'year'}).
}
\examples{
dat <- data.frame(
  bay_seg = rep(1:2, each = 6),
  basin = rep(c("02304500", "02306647"), each = 6),
  yr = rep(2021:2022, each = 3, times = 2),
  mo = rep(1:3, times = 4),
  clucsid = rep(1:3, times = 4),
  tnload = c(150, 250, 50, 180, 300, 40, 160, 270, 45, 170, 280, 35),
  tpload = c(15, 35, 8, 18, 42, 6, 16, 38, 7, 17, 40, 5),
  tssload = c(1200, 3500, 400, 1400, 4000, 350, 1300, 3800, 380, 1350, 3900, 320),
  bodload = c(800, 1500, 200, 900, 1800, 180, 850, 1600, 190, 870, 1650, 170),
  h2oload = c(50000, 80000, 25000, 55000, 85000, 22000, 52000, 82000, 23000, 53000, 83000, 21000)
)

util_nps_lusumm(dat, summ = 'basin', summtime = 'month')
}
