% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nps_preplog.R
\name{util_nps_preplog}
\alias{util_nps_preplog}
\title{Utility function for non-point source (NPS) ungaged workflow to prepare land use and soil data for logistic regression}
\usage{
util_nps_preplog(tbbase)
}
\arguments{
\item{tbbase}{Input data frame returned from \code{\link{util_nps_tbbase}}.}
}
\value{
A data frame with land use and soil areas in long format with bay segment and basins in the rows.
}
\description{
Utility function for non-point source (NPS) ungaged workflow to prepare land use and soil data for logistic regression
}
\examples{
util_nps_preplog(tbbase)
}
