% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nps_preprain.R
\name{util_nps_preprain}
\alias{util_nps_preprain}
\title{Prep rain data for non-point source (NPS) ungaged load estimates}
\usage{
util_nps_preprain(rain, yrrng = NULL)
}
\arguments{
\item{rain}{data frame of rainfall data, see \code{\link{rain}}}

\item{yrrng}{numeric vector of a single year or a two year range to filter the data, defaults to NULL (no filtering)}
}
\value{
A data frame of monthly total rainfall (inches) by drainage basin
}
\description{
Prep rain data for non-point source (NPS) ungaged load estimates
}
\details{
The inverse distance weighting scheme of the drainage basin to the rain gauge is used (with \code{\link{nps_distance}}) to estimate a cumulative monthly total rainfall, including lags of one and two months.
}
\examples{
util_nps_preprain(rain)

util_nps_preprain(rain, yrrng = c(2021))
}
