% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nps_segment.R
\name{util_nps_segment}
\alias{util_nps_segment}
\title{Create bay segment column for non-point source (NPS) load data}
\usage{
util_nps_segment(dat)
}
\arguments{
\item{dat}{data frame with a \code{basin} and \code{bay_seg} columns}
}
\value{
The same data frame with an additional \code{segment} column indicating the major bay segment associated with each row
}
\description{
Create bay segment column for non-point source (NPS) load data
}
\details{
This is a simple helper function used internally with \code{\link{anlz_nps}} and \code{\link{util_nps_lusumm}} to create a \code{segment} column based on the \code{basin} and \code{bay_seg} columns.
}
\examples{
dat <- data.frame(
  basin = c("LTARPON", "TBYPASS", "02300500", "206-4", "EVERSRES", "UNKNOWN"),
  bay_seg = c(1, 2, 3, 4, 7, NA)
)
util_nps_segment(dat)
}
