% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nps_unionchunk.R
\name{util_nps_unionchunk}
\alias{util_nps_unionchunk}
\title{Helper function for union operation}
\usage{
util_nps_unionchunk(sf1, sf2, chunk_size)
}
\arguments{
\item{sf1}{First sf object}

\item{sf2}{Second sf object}

\item{chunk_size}{Integer. For large datasets, process in chunks of this many features from sf1.}
}
\value{
An sf object containing the spatial intersection of sf1 and sf2, with geometries unioned by unique combinations of all attributes from both input objects.
}
\description{
Helper function for union operation
}
\details{
Used internally by \code{\link{util_nps_union}}. See the help file for more details.
}
