% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nps_unionnochunk.R
\name{util_nps_unionnochunk}
\alias{util_nps_unionnochunk}
\title{Helper function for union operation}
\usage{
util_nps_unionnochunk(sf1, sf2)
}
\arguments{
\item{sf1}{First sf object}

\item{sf2}{Second sf object}
}
\value{
An sf object containing the spatial intersection of sf1 and sf2, with geometries unioned by unique combinations of all attributes from both input objects.
}
\description{
Helper function for union operation
}
\details{
Used internally by \code{\link{util_nps_union}}. See the help file for more details.
}
\examples{
\dontrun{
data(tbjuris)
data(tbsubshed)
result <- util_nps_unionnochunk(tbsubshed, tbjuris)
}
}
