% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ps_fillmis.R
\name{util_ps_fillmis}
\alias{util_ps_fillmis}
\title{Fill missing point source data with annual average}
\usage{
util_ps_fillmis(dat)
}
\arguments{
\item{dat}{data frame from raw entity data as \code{data.frame}}
}
\value{
Input data frame as is if no missing values, otherwise missing data filled as described above.
}
\description{
Fill missing point source data with annual average
}
\details{
Missing concentration data are replaced with the average for the outfall in a given year. All flow data are also floored at zero.  Rows with missing flow data are assigned 0 for all data.  Rows with zero flow are assigned concentration of zero.
}
\examples{
pth <- system.file('extdata/ps_dom_hillsco_falkenburg_2019.txt', package = 'tbeploads')
dat <- read.table(pth, skip = 0, sep = '\t', header = TRUE)
dat <- util_ps_checkuni(dat)
util_ps_fillmis(dat)
}
