% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ps_fixoutfall.R
\name{util_ps_fixoutfall}
\alias{util_ps_fixoutfall}
\title{Light edits to the outfall ID column for point source data}
\usage{
util_ps_fixoutfall(dat)
}
\arguments{
\item{dat}{data frame from raw entity data as \code{data.frame}}
}
\value{
Input data frame as is, with any edits to the outfall ID column.
}
\description{
Light edits to the outfall ID column for point source data
}
\details{
The outfall ID column is edited lightly to remove any leading or trailing white space, a hyphen is added between letters and numbers if missing, and "Outfall" prefix is removed if presenn.
}
\examples{
pth <- system.file('extdata/ps_ind_busch_busch_2020.txt', package = 'tbeploads')
dat <- read.table(pth, skip = 0, sep = '\t', header = TRUE)
util_ps_fixoutfall(dat)
}
