% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ps_summ.R
\name{util_ps_summ}
\alias{util_ps_summ}
\title{Summarize point source load estimates}
\usage{
util_ps_summ(
  dat,
  summ = c("entity", "facility", "segment", "all"),
  summtime = c("month", "year")
)
}
\arguments{
\item{dat}{Pre-processed data frame of point source load estimates, see examples}

\item{summ}{chr string indicating how the returned data are summarized, see details}

\item{summtime}{chr string indicating how the returned data are summarized temporally (month or year), see details}
}
\value{
Data frame with summarized loading data based on user-supplied arguments
}
\description{
Summarize point source load estimates
}
\details{
The data are summarized differently based on the \code{summ} and \code{summtime} arguments.  All loading data are summed based on these arguments, e.g., by bay segment (\code{summ = 'segment'}) and year (\code{summtime = 'year'}).
}
\examples{
fls <- list.files(system.file('extdata/', package = 'tbeploads'),
  pattern = 'ps_ind_', full.names = TRUE)

ipsbyfac <- anlz_ips_facility(fls)

# add bay segment and source, there should only b loads to hills, middle, and lower tampa bay
ipsld <- ipsbyfac  |>
  dplyr::arrange(coastco) |>
  dplyr::left_join(dbasing, by = "coastco") |>
  dplyr::mutate(
    segment = dplyr::case_when(
      bayseg == 1 ~ "Old Tampa Bay",
      bayseg == 2 ~ "Hillsborough Bay",
      bayseg == 3 ~ "Middle Tampa Bay",
      bayseg == 4 ~ "Lower Tampa Bay",
      TRUE ~ NA_character_
    ),
    source = 'IPS'
  ) |>
  dplyr::select(-basin, -hectare, -coastco, -name, -bayseg)

util_ps_summ(ipsld, summ = 'entity', summtime = 'year')
}
