% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_ml.R
\name{anlz_ml}
\alias{anlz_ml}
\title{Calculate material loss (ML) loads and summarize}
\usage{
anlz_ml(
  fls,
  summ = c("entity", "facility", "segment", "all"),
  summtime = c("month", "year")
)
}
\arguments{
\item{fls}{vector of file paths to raw entity data, one to many}

\item{summ}{chr string indicating how the returned data are summarized, see details}

\item{summtime}{chr string indicating how the returned data are summarized temporally (month or year), see details}
}
\value{
data frame with loading data for TN as tons per month/year.  Columns for TP, TSS, BOD, and hydrologic load are also returned with zero load for consistency with other point source load calculation functions.
}
\description{
Calculate material loss (ML) loads and summarize
}
\details{
Input data files in \code{fls} are first processed by \code{\link{anlz_ml_facility}} to calculate ML loads for each facility.  The data are summarized differently based on the \code{summ} and \code{summtime} arguments.  All loading data are summed based on these arguments, e.g., by bay segment (\code{summ = 'segment'}) and year (\code{summtime = 'year'}). Options for \code{summ} are 'entity' to summarize by entity only, 'facility' to summarize by facility only, 'segment' to summarize by bay segment, and 'all' to summarize total load.  Options for \code{summtime} are 'month' to summarize by month and 'year' to summarize by year.  The default is to summarize by entity and month.
}
\examples{
fls <- list.files(system.file('extdata/', package = 'tbeploads'),
  pattern = 'ps_indml', full.names = TRUE)
anlz_ml(fls)
}
\seealso{
\code{\link{anlz_ml_facility}}
}
