% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_ml_facility.R
\name{anlz_ml_facility}
\alias{anlz_ml_facility}
\title{Calculate material losses (ML) from raw facility data}
\usage{
anlz_ml_facility(fls)
}
\arguments{
\item{fls}{vector of file paths to raw facility data, one to many}
}
\value{
data frame that is nearly identical to the input data except results are shown as monthly load as the annual estimate divided by 12.  This is for consistency of reporting without sources.
}
\description{
Calculate material losses (ML) from raw facility data
}
\details{
Input data should one row per year per facility, where the row shows the total tons per year of total nitrogen loss.
}
\examples{
fls <- list.files(system.file('extdata/', package = 'tbeploads'),
  pattern = 'ps_indml', full.names = TRUE)
anlz_ml_facility(fls)
}
\seealso{
\code{\link{anlz_ml}}
}
