% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_nps_gaged.R
\name{anlz_nps_gaged}
\alias{anlz_nps_gaged}
\title{Calculate non-point source (NPS) loads for gaged basins}
\usage{
anlz_nps_gaged(
  yrrng = c("2021-01-01", "2023-12-31"),
  mancopth = NULL,
  pincopth = NULL,
  lakemanpth = NULL,
  tampabypth = NULL,
  bellshlpth = NULL,
  allflo = NULL,
  allwq = NULL,
  usgsflow = NULL,
  verbose = TRUE
)
}
\arguments{
\item{yrrng}{A vector of two dates in 'YYYY-MM-DD' format, specifying the date range to retrieve flow data. Default is from '2021-01-01' to '2023-12-31'.}

\item{mancopth}{character, path to the Manatee County water quality data file, see details}

\item{pincopth}{character, path to the Pinellas County water quality data file, see details}

\item{lakemanpth}{character, path to the file containing the Lake Manatee flow data}

\item{tampabypth}{character, path to the file containing the Tampa Bypass flow data}

\item{bellshlpth}{character, path to the file containing the Bell shoals data}

\item{allflo}{data frame of flow data, if already available from \code{\link{util_nps_getflow}}, otherwise NULL and the function will retrieve the data}

\item{allwq}{data frame of water quality data, if already available from \code{\link{util_nps_getwq}}, otherwise NULL and the function will retrieve the data.}

\item{usgsflow}{data frame of USGS flow data, if already available from \code{\link{util_nps_getusgsflow}}, otherwise NULL and the function will retrieve the data. Default is NULL. Does not apply if \code{allflo} is provided.}

\item{verbose}{logical indicating whether to print verbose output}
}
\value{
A data frame with columns for basin, year, month, TN in mg/L, TP in mg/L, TSS in mg/L, BOD in mg/L, flow in liters/month, hydrologic load in m3/month, TN load in kg/month, TP load in kg/month, TSS load in kg/month, and BOD load in kg/month.
}
\description{
Calculate non-point source (NPS) loads for gaged basins
}
\details{
The function uses \code{\link{util_nps_getflow}} to retrieve flow data and \code{\link{util_nps_getwq}} to retrieve water quality data. It then combines these datasets and calculates loads for TN, TP, TSS, BOD, and hydrologic load.  See the help files for each function for more details.

Required external data inputs are Lake Manatee, Tampa Bypass, and Alafia River Bell Shoals flow data.  These are not available from the USGS API and must be obtained from the contacts listed in \code{\link{util_nps_getextflow}}.  USGS flow data are for stations 02299950, 02300042, 02300500, 02300700, 02301000, 02301300, 02301500, 02301750, 02303000, 02303330, 02304500, 02306647, 02307000, 02307359, and 02307498.  The USGS flow data are from the NWIS database as returned by \code{\link[dataRetrieval]{read_waterdata_daily}} using \code{\link{util_nps_getusgsflow}}. A preprocessed USGS flow data frame can be provided using the \code{usgsflow} argument to avoid re-downloading the data.  All inputs for flow can be superseded by providing a complete flow data frame using the \code{allflo} argument.

Water Quality data are obtained from the FDEP WIN database API, tbeptools, or local files as described in \code{\link{util_nps_getwq}}. Chosen stations are ER2 and UM2 for Manatee County and station 06-06 for Pinellas County. Environmental Protection Commission (EPC) of Hillsborough County stations retained are 105, 113, 114, 132, 141, 138, 142, and 147. Manatee or Pinellas County data can be imported from local files using the \code{mancopth} and \code{pincopth} arguments, respectively.  If these are not provided, the function will attempt to retrieve data from the FDEP WIN database using \code{read_importwqwin} from tbeptools.  The EPC data are retrieved using \code{read_importepc} from tbeptools.  All inputs for water quality can be superseded by providing a complete water quality data frame using the \code{allwq} argument.

The function assumes that the water quality data are in mg/L and flow data are in cfs.  Missing water quality data are filled with previous five year averages for the end months, then linearly interpolated using \code{\link{util_nps_fillmiswq}}.
}
\examples{
data(allwq)
data(allflo)

nps_gaged <- anlz_nps_gaged(
  yrrng = c('2021-01-01', '2023-12-31'), 
  allflo = allflo,
  allwq = allwq
)

head(nps_gaged)
}
