% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nps_fillwq.R
\name{util_nps_fillmiswq}
\alias{util_nps_fillmiswq}
\title{Fill in missing water quality values for non-point source (NPS) data}
\usage{
util_nps_fillmiswq(wq)
}
\arguments{
\item{wq}{A data frame of water quality data returned by \code{util_nps_getwq}}
}
\value{
Input data frame with missing data filled as described above.
}
\description{
Fill in missing water quality values for non-point source (NPS) data
}
\details{
Missing end date monthly values are filled with prior 5 year averages. Then, missing monthly values are linearly interpolated using \code{\link[zoo]{na.approx}}.  This function will need to be updated each RA period with correct averages.
}
\examples{
\dontrun{
manopth <- system.file('extdata/nps_wq_manco.txt', package = 'tbeploads')
pincopth <- system.file('extdata/nps_wq_pinco.txt', package = 'tbeploads')
wq <- util_nps_getwq(c('2021-01-01', '2023-12-31'), mancopth = manopth,
  pincopth = pincopth, verbose = F)
wq <- util_nps_fillmiswq(wq)
}
}
