% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nps_getflow.R
\name{util_nps_getflow}
\alias{util_nps_getflow}
\title{Get flow data from for NPS calculations at gaged sites}
\usage{
util_nps_getflow(
  lakemanpth,
  tampabypth,
  bellshlpth,
  yrrng = c(2021, 2023),
  usgsflow = NULL,
  verbose = TRUE
)
}
\arguments{
\item{lakemanpth}{character, path to the file containing the Lake Manatee flow data}

\item{tampabypth}{character, path to the file containing the Tampa Bypass flow data}

\item{bellshlpth}{character, path to the file containing the Bell shoals data}

\item{yrrng}{vector of two integers, the year range for which to retrieve flow data. Default is c(2021, 2023).}

\item{usgsflow}{data frame of USGS flow data, if already available from \code{\link{util_nps_getusgsflow}}, otherwise NULL and the function will retrieve the data. Default is NULL.}

\item{verbose}{logical indicating whether to print verbose output}
}
\value{
A data frame of monthly mean flow for fifteen USGS stations and three external flow sites
}
\description{
Get flow data from for NPS calculations at gaged sites
}
\details{
Missing flow values are linearly interpolated using \code{\link[zoo]{na.approx}}.  The function combines external and USGS API flow data using the \code{util_nps_getextflow} and \code{util_nps_getusgsflow} functions.

A preprocessed USGS flow data frame can be provided using the \code{usgsflow} argument to avoid re-downloading the data.
}
\examples{
lakemanpth <- system.file('extdata/nps_extflow_lakemanatee.xlsx', package = 'tbeploads')
tampabypth <- system.file('extdata/nps_extflow_tampabypass.xlsx', package = 'tbeploads')
bellshlpth <- system.file('extdata/nps_extflow_bellshoals.xls', package = 'tbeploads')
allflo <- util_nps_getflow(lakemanpth, tampabypth, bellshlpth, usgsflow = usgsflow)
}
\seealso{
\code{\link{util_nps_getextflow}}, \code{\link{util_nps_getusgsflow}}
}
