% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nps_getwq.R
\name{util_nps_getwq}
\alias{util_nps_getwq}
\title{Get water quality data for NPS gaged flows}
\usage{
util_nps_getwq(yrrng = c("2021-01-01", "2023-12-31"), verbose = TRUE)
}
\arguments{
\item{yrrng}{A vector of two dates in 'YYYY-MM-DD' format, specifying the date range to retrieve flow data. Default is from '2021-01-01' to '2023-12-31'.}

\item{verbose}{logical indicating whether to print verbose output}
}
\value{
A data frame of water quality data for Manatee, Pinellas, and Hillsborough County
}
\description{
Get water quality data for NPS gaged flows
}
\details{
Manatee and Pinellas County data is retrieved from the FDEP WIN database using \code{\link[tbeptools]{read_importwqwin}}.  Hillsborough County data is retrieved using \code{\link[tbeptools]{read_importwq}}.
}
\examples{
\dontrun{
wqdat <- util_nps_getwq(c('2021-01-01', '2023-12-31'))
}
}
